% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePairwiseSimilarities.R
\name{calculatePairwiseSimilarities}
\alias{calculatePairwiseSimilarities}
\title{Run plink to Calculate Pairwise Similarities}
\usage{
calculatePairwiseSimilarities(
  file,
  db = NA,
  map = NA,
  out = NA,
  sexing = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to the filtered ped/map file pair (without ped/map file extension)}

\item{db}{Path to an existing genotype database}

\item{map}{Filepath to PlateDnoY.map file}

\item{out}{Path to the output, can be emptu}

\item{verbose}{Should the output be verbose, logical or numerical}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function is a wrapper to plink to calculate pairwise similarities
}
