% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateErrors.R
\name{estimateErrors}
\alias{estimateErrors}
\title{Process the PLINK ped files}
\usage{
estimateErrors(
  file,
  db = NA,
  appendSamplesToDB = FALSE,
  keep.rep = 1,
  y.marker = NA,
  x.marker = NA,
  plots = TRUE,
  neg_controls = NA,
  allele_error = 5,
  marker_dropout = 15,
  no_marker = 50,
  male.y = 3,
  male.hetX = 0,
  female.y = 0,
  female.Xtot = 8,
  female.hetXtot = 3,
  warning.noYtot = 2,
  warning.noHetXtot = 3,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to the ped input file}

\item{keep.rep}{numeric, keep only this n-fold replicates, default n=2}

\item{y.marker}{logical, Remove markers located on y-chromosome}

\item{plots}{logical, shall plots be created?}

\item{verbose}{Should the output be verbose, logical or numerical}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function processes the PLINK ped files
}
