% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidigm2PLINK.R
\name{fluidigm2PLINK}
\alias{fluidigm2PLINK}
\title{Turn fluidigm into PLINK format}
\usage{
fluidigm2PLINK(
  file,
  out = NA,
  ymap = "PlateD_withY.map",
  verbose = TRUE,
  plots = TRUE
)
}
\arguments{
\item{file}{Path to the input file}

\item{out}{Out file name, keep empty to keep the original basename}

\item{ymap}{Filepath to PlateD_withY.map file}

\item{verbose}{Should the output be verbose, logical}
}
\value{
Something
}
\description{
This function loads a fluidigm output and turns it into a PLINK format
}
