% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidigmAnalysisWrapper.R
\name{fluidigmAnalysisWrapper}
\alias{fluidigmAnalysisWrapper}
\title{Run the fluidigm analysis script together}
\usage{
fluidigmAnalysisWrapper(
  file,
  out = NA,
  db,
  ymap = "PlateD_withY.map",
  woymap,
  keep.rep = 2,
  remove.y = TRUE,
  plots = TRUE,
  rearrange = FALSE,
  group = NA,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to the fluidigm input file}

\item{out}{Out file name, keep empty to keep the original basename}

\item{db}{Filepath to database file}

\item{ymap}{Filepath to PlateD_withoutY.map file}

\item{keep.rep}{numeric, keep only this n-fold replicates, default n=2}

\item{remove.y}{logical, Remove markers located on y-chromosome}

\item{plots}{logical, shall plots be created?}

\item{verbose}{Should the output be verbose, logical or numerical}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function is a wrapper for the whole analysis
}
