% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairwiseSimilarityLoci.R
\name{getPairwiseSimilarityLoci}
\alias{getPairwiseSimilarityLoci}
\title{Get Pairwise Similarity Loci}
\usage{
getPairwiseSimilarityLoci(file, verbose = TRUE)
}
\arguments{
\item{file}{Path to the previous plink created ped file (with ped extension)}

\item{verbose}{Should the output be verbose, logical or numerical}

\item{db}{Path to an existing genotype database}

\item{map}{Filepath to PlateDnoY.map file}

\item{out}{Path to the output, can be emptu}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function is a wrapper to a perl script to determine the loci of pairwise similarities
}
