% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateErrors.R
\name{estimateErrors}
\alias{estimateErrors}
\title{Process the PLINK ped files}
\usage{
estimateErrors(
  file,
  keep.rep = 2,
  remove.y = TRUE,
  plots = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to the ped input file}

\item{keep.rep}{numeric, keep only this n-fold replicates, default n=2}

\item{remove.y}{logical, Remove markers located on y-chromosome}

\item{plots}{logical, shall plots be created?}

\item{verbose}{Should the output be verbose, logical or numerical}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function processes the PLINK ped files
}
