% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fluidigm-package.R
\docType{package}
\name{Fluidigm}
\alias{Fluidigm}
\alias{Fluidigm-package}
\title{Fluidigm}
\description{
The package is aimed for Fluidigm users to handle data coming in Plink format.
}
\section{Title}{

A brief one-line description of the package.
}

\section{Description}{

tba
}

\section{Details}{

tba
}

\section{References}{

tba
}

\author{
\strong{Maintainer}: Daniel Fischer \email{Daniel.Fischer@luke.fi} (\href{https://orcid.org/0000-0002-7513-683X}{ORCID})

Authors:
\itemize{
  \item Robert Ekblom (\href{https://orcid.org/0000-0003-2222-1966}{ORCID})
}

Other contributors:
\itemize{
  \item Helena Johansson [contributor, funder]
}

}
\keyword{fluidigm}
\keyword{genotyping}
