% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidigm2PLINK.R
\name{fluidigm2PLINK}
\alias{fluidigm2PLINK}
\title{Turn Fluidigm Output Into PLINK Format}
\usage{
fluidigm2PLINK(
  file,
  out = NA,
  map = NA,
  plots = TRUE,
  rearrange = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{Path to the input file}

\item{out}{Out file name, keep empty to keep the original basename}

\item{map}{Filepath to PlateD_withY.map file}

\item{rearrange}{Logical, rearrange the ped/map output in order of ymap}

\item{verbose}{Should the output be verbose, logical or numerical}

\item{plot}{Logical, plot additional figures for conversion}
}
\value{
A ped/map file pair and optional diagnostic plots
}
\description{
This function loads a fluidigm output and turns it into a PLINK format
}
