% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairwiseSimilarityLoci.R
\name{getPairwiseSimilarityLoci}
\alias{getPairwiseSimilarityLoci}
\title{Get Pairwise Similarity Loci}
\usage{
getPairwiseSimilarityLoci(file, verbose = TRUE, verbosity = 1)
}
\arguments{
\item{file}{Path to the previous plink created ped file (with ped extension)}

\item{verbose}{Should the output be verbose, logical}

\item{verbosity}{Level of verbosity, set to higher number for more details

#' @details
Additional details...}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function is a wrapper to a perl script to determine the loci of pairwise similarities
}
\examples{
\dontrun{
  getPairwiseSimilarityLoci()
}

}
