% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMatrix.R
\name{similarityMatrix}
\alias{similarityMatrix}
\title{Run the fluidigm analysis script together}
\usage{
similarityMatrix(
  file = NA,
  mibs.file = NA,
  pairs.file = NA,
  ped.file = NA,
  group = NA,
  plots = TRUE,
  similarity = 0.85,
  verbose = TRUE,
  verbosity = 1
)
}
\arguments{
\item{file}{Path to the input file}

\item{mibs.file}{Path to the mibs input file (only if name differs)}

\item{pairs.file}{Path to the pairs input file (only if name differs)}

\item{ped.file}{Path to the pairs input file (only if name differs)}

\item{group}{Sample identified for sample-wise statistics}

\item{plots}{logical, shall plots be created?}

\item{similarity}{Threshold defining the level of similarity}

\item{verbose}{Should the output be verbose, logical}

\item{verbosity}{Level of verbosity, set to higher number for more details

#' @details
Additional details...}
}
\value{
A list containing the following elements:
gensim, a matrix indicating if genotypes are called correctly for replicates and/or if genotypes are missing
summs, a matrix with summary statistics
}
\description{
This function is a wrapper for the whole analysis
}
\examples{
\dontrun{
  similarityMatrix()
}

}
